fix free_space problem

+++ src/includes.cpp
@@ -73,12 +73,12 @@ unsigned int get_lang_color(std::string language) {
     return RGBA8(36, 41, 46,255);
 }
 
-float byte_to_kb(int bytes) {
+float byte_to_kb(uint64_t bytes) {
     return (float)bytes/1024;
 }
-float byte_to_mb(int bytes) {
+float byte_to_mb(uint64_t bytes) {
     return (float)bytes/1024/1024;
 }
-float byte_to_gb(int bytes) {
+float byte_to_gb(uint64_t bytes) {
     return (float)bytes/1024/1024/1024;
 }
